Выводятся ошибки сообщений при неправильных подписках на пользователей. Решена проблема со статус. Кто-то дура забыла импортировать Response

+++ backend/foodgram/users/views.py
@@ -1,13 +1,14 @@
-from django.shortcuts import get_object_or_404, render
-from djoser.views import UserViewSet
 from requests import Response
 from rest_framework import status
+
+from django.shortcuts import get_object_or_404
+from djoser.views import UserViewSet
+from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework.pagination import PageNumberPagination
-from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
-from api.permissions import AuthorOrReadOnly
+from rest_framework.permissions import IsAuthenticated
+
 from .models import CustomUser, Subscribed
-from recipes.serializers import SubscribedSerializer
 from .serializers import UserCreateSerializer, UserSerializer
 from recipes.serializers import SubscribedSerializer
 
@@ -43,18 +44,24 @@ class UserViewSet(UserViewSet):
             serializer = SubscribedSerializer(author,
                                               context={'request': request})
             Subscribed.objects.create(user=user, author=author)
-            return Response(serializer.data, {'message': 'Вы подписались на автора'}, status=status.HTTP_201_CREATED)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         if request.method == 'DELETE':
-            subscript = get_object_or_404(Subscribed, user=user, author=author)
-            subscript.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if Subscribed.objects.filter(user=user, author=author).exists():
+                subscript = get_object_or_404(Subscribed, user=user, author=author)
+                subscript.delete()
+                return Response(
+                    {'message': 'Вы отписались от автора'},
+                    status=status.HTTP_204_NO_CONTENT)
+            return Response(
+                {'error': 'Вы не подписаны на данного автора'},
+                status=status.HTTP_400_BAD_REQUEST)
         return Response(status)
 
     @action(detail=False,
             permission_classes=[IsAuthenticated],
             methods=('GET',))
     def subscriptions(self, request):
-        user = self.request.user
+        user = request.user
         authors = CustomUser.objects.filter(subscribing__user=user)
         serializer = SubscribedSerializer(self.paginate_queryset(authors),
                                           many=True,
